Fix mapping out of bounds error.
When removing a mapping the previous code assumed the number of active assigned mappings was always equal to the size of the mapping array. This is not the case as we do not erase elements when we remove mappings due to needing to preserve version info.

This bugfix added a seperate count to the mappings to now track the number of active mappings appropriatly.