Modify World.DestroyEntity() to correctly return the entity handle to the pool and modify the destroy system to destroy one by one.
The previous optimization to destroy all entities within an archetype was not able to correctly return Entities back to their allocated pool so instead I moved to the less optimal procedure of deleting each entity one by one. I may move the task of allocating/deallocating entities into the archetypes to enable this optimization to be put back in the future.